+++ lib/wardrobe/plugins/validation/instance_methods.rb
@@ -5,32 +5,32 @@ module Wardrobe
     module Validation
       module InstanceMethods
         def _validate
+          return ValidationRunner.validate(self) if frozen?
           unless _validated?
             @_validator = ValidationRunner.validate(self)
             @_validated = true
           end
-          self
+          @_validator
         end
 
         def _validate!
-          _validate unless _validated?
-          raise ValidationError.new(@_validator.errors) unless _valid?
+          validator = _validate
+          raise ValidationError.new(validator.errors) unless validator.errors.empty?
           self
         end
 
         def _valid?
-          _validate unless _validated?
-          @_validator.errors.empty?
+          _validate.errors.empty?
         end
 
         def _validation_errors
-          _validate unless _validated?
-          @_validator.errors
+          _validate.errors
         end
 
         private
 
         def _validated?
+          return false if frozen?
           @_validated ||= false
         end
       end